Timing: factor out timing, eb-console and picoscope and make optional
